Fix the piechartwidget, calculate the percents with double instead of int, prevent data lost
svn path=/trunk/KDE/kdenetwork/kget/; revision=771320

+++ plasma/applet/piechartwidget.cpp
@@ -204,7 +204,7 @@ public:
     QSize size;
     QMap <QString, PrivateData> data;
     int bottomMargin;
-    float totalSize;
+    double totalSize;
     bool emitUpdateGeometrySignal;
 };
 
@@ -219,24 +219,24 @@ PieChartWidget::~PieChartWidget()
     delete d;
 }
 
-void PieChartWidget::addData(const QString &name, int length)
+void PieChartWidget::addData(const QString &name, double length)
 {
     addData(name, length, 0, false);
 }
 
-void PieChartWidget::addData(const QString &name, int length, int activeLength)
+void PieChartWidget::addData(const QString &name, double length, double activeLength)
 {
     addData(name, length, activeLength, false);
 }
 
-void PieChartWidget::addData(const QString &name, int length, int activeLength, bool active)
+void PieChartWidget::addData(const QString &name, double length, double activeLength, bool active)
 {
     if (!d->data.contains(name)) {
-        d->data [name] = PrivateData ();
+        d->data [name] = PrivateData();
 
         // update the widget size
         d->bottomMargin += 20;
-        d->size.setHeight (d->size.height() + 20);
+        d->size.setHeight(d->size.height() + 20);
 
         d->emitUpdateGeometrySignal = true;
     }
